Initializes PlotFragment

+++ app/src/main/java/pl/lodz/mobile/covidinfo/utility/AndroidUtils.kt
@@ -5,7 +5,7 @@ import android.util.DisplayMetrics
 import kotlin.math.roundToInt
 
 
-fun dpToPixels(context: Context, dp: Int): Int {
+fun dpToPixels(context: Context, dp: Float): Int {
     val displayMetrics: DisplayMetrics = context.resources.displayMetrics
     return (dp * (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT)).roundToInt()
 }
